reverse archive file and unarchive file contents
